chore: Update .gitignore and vite.config.js
The .gitignore file was updated to include the "*.local" and ".vite" files. This ensures that these files are ignored and not tracked by Git.

In vite.config.js, the "base" property was added to the Vite configuration. It sets the base URL for the application based on the environment. In development, the base URL is set to an empty string, while in production, it is set to "/map-tiles-view/".

These changes are considered as a chore and do not introduce any new features or fix any bugs.